added more services to database service

+++ DBService/src/main/java/com/naga/spring/dbservice/model/Product.java
@@ -8,7 +8,7 @@ public class Product {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int productId;
+    private long productId;
 
     @Column
     private String productName;
@@ -23,18 +23,18 @@ public class Product {
 
     }
 
-    public Product(int productId, String productName, String productDescription, double productPrice) {
+    public Product(long productId, String productName, String productDescription, double productPrice) {
         this.productId = productId;
         this.productName = productName;
         this.productDescription = productDescription;
         this.productPrice = productPrice;
     }
 
-    public int getProductId() {
+    public long getProductId() {
         return productId;
     }
 
-    public void setProductId(int productId) {
+    public void setProductId(long productId) {
         this.productId = productId;
     }
